# swagger.json
{
  "info": {
    "title": "Blackstone",
    "version": "0.7.0",
    "description": "The Blackstone codebase is a collection of smart contracts and RESTful APIs which together form the basis for the Agreements Network. This collection includes a full-feature business process execution engine, and high-level translator from BPMN standard to smart contract based process graphs. It also include an object management suite utilized by the Agreements Network."
  },
  "swagger": "2.0",
  "paths": {
...
    "/archetypes": {
      "get": {
        "tags": [
          "Archetypes"
        ],
        "description": "archetype information. Within the Agreements Network, Archetypes are the fundamental, top level objects. They are holders for a set of information which allows users to creat Active Agreements within the Platform. The returned list will include archetypes that are: a. authored by the authenticated user, or b. public (ie. `isPrivate` property is `false`) and activated (ie. `active` property is true)",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "json",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "address": {
                    "type": "string",
                    "description": "Archetype's address"
                  },
                  "name": {
                    "type": "string",
                    "description": "Human readable name of the Archetype"
                  },
                  "author": {
                    "type": "string",
                    "description": "Controller contract of the user or organization\nthat created the Archetype"
                  },
                  "description": {
                    "type": "string",
                    "description": "description of the Archetype"
                  },
                  "active": {
                    "type": "boolean",
                    "description": "Whether the Archetype can be used to create\nnew Active Agreements or not"
                  },
                  "isPrivate": {
                    "type": "boolean",
                    "description": "Whether the encryption framework of the Archetype\nis operational or not"
                  },
                  "numberOfParameters": {
                    "type": "number",
                    "description": "The number of custom parameters used by the Archetype"
                  },
                  "numberOfDocuments": {
                    "type": "number",
                    "description": "The number of documents registered against the Archetype"
                  },
                  "countries": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "The jurisdictions in which the Archetype has been\nregistered to be active"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Archetypes"
        ],
        "description": "Create an Archetype",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "description": "Create an Archetype",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "description": "Human readable name of the Archetype (limit: 32 ASCII characters)"
                },
                "author": {
                  "type": "string",
                  "description": "Controller contract of the user or organization\nthat created the Archetype. Note that currently the author is forced to be the logged-in-user's address\nsince we do not yet have a review/approval process for changes made to an organization-authored archetype by a user who is part of that organization"
                },
                "description": {
                  "type": "string",
                  "description": "Short human readable description of the Archetype"
                },
                "price": {
                  "type": "integer",
                  "description": "Price of the archetype, in cents (Optional- this field can be edited later through a `PUT` request to '/archetypes/:address/price')"
                },
                "isPrivate": {
                  "type": "boolean",
                  "description": "**(Optional)** Whether the encryption framework of\nthe Archetype is operational or not"
                },
                "password": {
                  "type": "string",
                  "description": "A secret string which is used to trigger the encryption\nsystem for the Archetype's documents"
                },
                "parameters": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "signatory": {
                        "type": "string"
                      }
                    }
                  },
                  "description": "**(Optional)** The \"name\" (limit: 32 ASCII characters) and \"type\" of all custom parameters used\nby the Archetype"
                },
                "documents": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "grant": {
                        "type": "string"
                      }
                    }
                  },
                  "description": "**(Optional)**  The \"name\" and \"grant\" (if any)\nsufficient to provide the information regarding the relevant documents associated with\nthe Archetype"
                },
                "jurisdictions": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "country": {
                        "type": "string"
                      },
                      "regions": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "description": "The \"country\" and \"regions\" which the Archetype\nhas been registered as relevant to. The \"country\" is registered as an ISO standard\ntwo character string and \"regions\" is an array of addresses relating to the controlling\ncontracts for the region (see [ISO standards manipulation](#) section)."
                },
                "formationProcessDefinition": {
                  "type": "string",
                  "description": "Address of the formation process definition\ncontroller"
                },
                "executionProcessDefinition": {
                  "type": "string",
                  "description": "Address of the execution process definition\ncontroller"
                },
                "governingArchetypes": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "description": "Array of contract addresses of the archetypes to govern this new one\nof the rights and conditions of the contract"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "json",
            "schema": {
              "type": "object",
              "properties": {
                "address": {
                  "type": "string",
                  "description": "address of the created Archetype"
                }
              }
            }
          }
        }
      }
    },
    "/archetype-packages/{id}": {
      "get": {
        "tags": [
          "Archetypes"
        ],
        "description": "Gets information for a single archetype package. Returns a `404` if the package is private or not active and the authenticated user is not its author.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "description": "archetype package id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "json",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "description": "Archetype Package id"
                },
                "name": {
                  "type": "string",
                  "description": "Human readable name of the Archetype Package"
                },
                "description": {
                  "type": "string",
                  "description": "Human readable description of the Archetype Package"
                },
                "author": {
                  "type": "string",
                  "description": "Controller contract of the user or organization"
                },
                "isPrivate": {
                  "type": "boolean",
                  "description": "Indicates whether the package can be read/used publicly"
                },
                "active": {
                  "type": "boolean",
                  "description": "Indicates whether the package has been activated and available for creating collections with."
                },
                "archetypes": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "address": {
                        "type": "string"
                      },
                      "active": {
                        "type": "boolean"
                      }
                    }
                  },
                  "description": "Array of archetypes with name, address, and active keys\nthat are included in the Archetype Package"
                }
              }
            }
          }
        }
      }
    },
    "/archetypes/packages": {
      "post": {
        "tags": [
          "Archetypes"
        ],
        "description": "Create an Archetype Package",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "description": "Create an Archetype Package",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "description": "Human readable name of the Archetype Package"
                },
                "description": {
                  "type": "string",
                  "description": "Human readable description of the Archetype Package"
                },
                "String": {
                  "type": "string",
                  "description": "author Controller contract of the user or organization. Will be set to the authenticated user's address if not provided.\nthat created the Archetype Package"
                },
                "isPrivate": {
                  "type": "boolean",
                  "description": "Human readable description of the Archetype Package"
                },
                "active": {
                  "type": "boolean",
                  "description": "Optional- will default to false"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "json",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "description": "id of the created Archetype Package"
                }
              }
            }
          }
        }
      }
    },
    "/archetype-packages/{id}/archetype/{address}": {
      "put": {
        "tags": [
          "Archetypes"
        ],
        "description": "Add an archetype to a package",
        "produces": [
          "text/plain"
        ],
        "parameters": [
          {
            "name": "address",
            "description": "Archetype address",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "description": "Package Id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Add an archetype to a package",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/archetype-packages/{id}/activate": {
      "put": {
        "tags": [
          "Archetypes"
        ],
        "description": "Activate the archetype package.\nAn archetype package can only be activated by its author.",
        "produces": [
          "text/plain"
        ],
        "parameters": [
          {
            "name": "id",
            "description": "Archetype package id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Activate an archetype package",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/archetype-packages/{id}/deactivate": {
      "put": {
        "tags": [
          "Archetypes"
        ],
        "description": "Deactivate the archetype package. An archetype package can only be deactivated by its author. Once an archetype package is deactivated by its author, it will not be included  in `GET /archetype-packges` or `GET /archetype-packages/:id` responses  made by users other than the author.",
        "produces": [
          "text/plain"
        ],
        "parameters": [
          {
            "name": "id",
            "description": "Archetype package id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "deactivate an archetype package",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
  },
...
  "securityDefinitions": {},
...
}